requirements.txt
pyperclip
PyPDF2
reportlab
python-docx
python-dotenv